rename: ModifyNickNameForm -> ModifyNickNameFormDTO
